WIP Clean up SketchExports
