+++ docs/src/components/SketchExports/SketchExports.less
@@ -1,10 +1,5 @@
 @import (reference) "~seek-style-guide/theme";
 
-.beforeReset::before,
-.afterReset::after {
-  content: none !important;
-}
-
 .text {
   margin-bottom: @grid-row-height;
 }
@@ -23,15 +18,6 @@
   margin-bottom: @grid-row-height;
 }
 
-.value {
-  .color & {
-    width: 100%;
-    border-radius: 3px;
-    height: @grid-row-height * 9;
-    margin-bottom: @grid-row-height * 2;
-  }
-}
-
 .color {
   @media @mobile {
     width: 100%;
@@ -45,4 +31,11 @@
 .colors {
   display: flex;
   flex-flow: row wrap;
+}
+
+.swatch {
+  width: 100%;
+  border-radius: 3px;
+  height: @grid-row-height * 9;
+  margin-bottom: @grid-row-height * 2;
 }
