+++ docs/src/components/SketchExports/fixSketchRendering/fixSketchRendering.less
@@ -1,0 +1,4 @@
+.beforeReset::before,
+.afterReset::after {
+  content: none !important;
+}
